requirements.txt
pandas
plotnine
python-dotenv
streamlit
tweepy